refactor(analysis): extract api surface symbol scanner
